# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Haber portalı için REST API",
        "title": "Haber API",
        "contact": {
            "name": "API Destek Ekibi",
            "email": "destek@haber.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/articles": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Yeni bir makale oluşturur (Sadece admin ve editörler)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Makaleler"
                ],
                "summary": "Yeni makale oluştur",
                "parameters": [
                    {
                        "description": "Makale bilgileri",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Article"
                        }
                    },
                    "400": {
                        "description": "Geçersiz istek",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Yetersiz yetki",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/articles/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mevcut bir makaleyi günceller (Sadece admin ve editörler)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Makaleler"
                ],
                "summary": "Makale güncelle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Makale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Makale bilgileri",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Article"
                        }
                    },
                    "400": {
                        "description": "Geçersiz istek",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Yetersiz yetki",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Makale bulunamadı",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Bir makaleyi siler (Sadece admin ve editörler)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Makaleler"
                ],
                "summary": "Makale sil",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Makale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Başarıyla silindi"
                    },
                    "400": {
                        "description": "Geçersiz istek",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Yetersiz yetki",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Makale bulunamadı",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/articles": {
            "get": {
                "description": "Tüm makaleleri sayfalanmış şekilde listeler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Makaleler"
                ],
                "summary": "Makaleleri listele",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sayfa numarası (varsayılan: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Sayfa başına sonuç sayısı (varsayılan: 10, maksimum: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Makale durumu (varsayılan: published)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Article"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/articles/author/{authorID}": {
            "get": {
                "description": "Belirli bir yazara ait makaleleri sayfalanmış şekilde listeler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Makaleler"
                ],
                "summary": "Yazara göre makaleleri getir",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Yazar ID",
                        "name": "authorID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Sayfa numarası (varsayılan: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Sayfa başına sonuç sayısı (varsayılan: 10, maksimum: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Article"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Geçersiz yazar ID",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/articles/category/{categoryID}": {
            "get": {
                "description": "Belirli bir kategoriye ait makaleleri sayfalanmış şekilde listeler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Makaleler"
                ],
                "summary": "Kategoriye göre makaleleri getir",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Kategori ID",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Sayfa numarası (varsayılan: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Sayfa başına sonuç sayısı (varsayılan: 10, maksimum: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Article"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Geçersiz kategori ID",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/articles/featured": {
            "get": {
                "description": "Öne çıkan makaleleri listeler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Makaleler"
                ],
                "summary": "Öne çıkan makaleleri getir",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maksimum makale sayısı (varsayılan: 5, maksimum: 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Article"
                            }
                        }
                    }
                }
            }
        },
        "/articles/slug/{slug}": {
            "get": {
                "description": "Slug'a göre makale detayı getirir",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Makaleler"
                ],
                "summary": "Slug ile makale getir",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Makale Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Article"
                        }
                    },
                    "404": {
                        "description": "Makale bulunamadı",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/articles/tag/{tagID}": {
            "get": {
                "description": "Belirli bir etikete sahip makaleleri sayfalanmış şekilde listeler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Makaleler"
                ],
                "summary": "Etikete göre makaleleri getir",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Etiket ID",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Sayfa numarası (varsayılan: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Sayfa başına sonuç sayısı (varsayılan: 10, maksimum: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Article"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Geçersiz etiket ID",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "description": "ID'ye göre makale detayı getirir",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Makaleler"
                ],
                "summary": "Makale detayı getir",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Makale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Article"
                        }
                    },
                    "400": {
                        "description": "Geçersiz makale ID",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Makale bulunamadı",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateArticleRequest": {
            "type": "object",
            "required": [
                "category_id",
                "content",
                "summary",
                "title"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "featured_image": {
                    "type": "string"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "published_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "published",
                        "draft",
                        "pending"
                    ]
                },
                "summary": {
                    "type": "string"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Hata açıklaması"
                },
                "message": {
                    "type": "string",
                    "example": "Detaylı hata mesajı"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "domain.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    }
}